# server/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: EduCenterPro API
  description: 'Tài liệu API đầy đủ cho EduCenterPro — nền tảng quản lý trung tâm ngoại ngữ.


    Gồm: Auth, Học viên, Giáo viên, Lớp học, Buổi học & Điểm danh, Học phí & Công nợ, Bài tập & Quiz, Sổ điểm, Phòng học,
    Lương, Học thử, Leads CRM, Referral, Reviews, Zalo OA, Cổng phụ huynh, Cổng giáo viên, Public landing.


    Tất cả endpoint staff đều yêu cầu JWT (đăng nhập admin/staff/teacher).

    Base path chính: /api/v1 (alias legacy /api vẫn hoạt động nhưng deprecated).'
  version: 1.0.0
security:
  - bearerAuth: []
servers:
  - url: http://localhost:4000
    description: Dev server
  - url: /
    description: Production (cùng origin)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Đăng nhập POST /api/auth/login để lấy token
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Không tìm thấy bài tập
        code:
          type: string
          example: NOT_FOUND
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 125
        totalPages:
          type: integer
          example: 7
    PaginatedStudents:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        pagination:
          $ref: '#/components/schemas/Pagination'
...
paths:
...
  /api/v1/students:
    get:
      tags:
        - Students
      summary: Danh sách học viên (filter + phân trang)
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Tìm theo tên/mã/SĐT
        - in: query
          name: status
          schema:
            type: string
            enum:
              - studying
              - paused
              - graduated
        - in: query
          name: class_id
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedStudents'
    post:
      tags:
        - Students
      summary: Thêm học viên mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Nguyễn Văn A
                phone:
                  type: string
                  example: 0901234567
                dob:
                  type: string
                  example: '2015-03-10'
                  description: YYYY-MM-DD
                class_ids:
                  type: array
                  items:
                    type: integer
                  description: Xếp lớp ngay
      responses:
        '201':
          description: Đã tạo
  /api/v1/students/{id}:
    get:
      tags:
        - Students
      summary: Chi tiết học viên
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '404':
          description: Không tìm thấy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Students
      summary: Cập nhật học viên
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                status:
                  type: string
                  enum:
                    - studying
                    - paused
                    - graduated
      responses:
        '200':
          description: OK
    delete:
      tags:
        - Students
      summary: Xóa học viên
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
...
  /api/v1/payroll:
    get:
      tags:
        - Payroll
      summary: Bảng lương giáo viên (tự tính từ điểm danh; tháng đã chốt trả bảng chụp lúc chốt)
      parameters:
        - in: query
          name: month
          schema:
            type: string
          description: YYYY-MM, mặc định tháng hiện tại
      responses:
        '200':
          description: OK
...
  /api/v1/trials:
    get:
      tags:
        - Trials
      summary: Danh sách học thử
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - new
              - contacted
              - converted
              - dropped
      responses:
        '200':
          description: OK
  /api/v1/trials/{id}:
    put:
      tags:
        - Trials
      summary: Cập nhật học thử
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/v1/trials/{id}/convert:
    post:
      tags:
        - Trials
      summary: Chuyển học thử thành học viên chính thức
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
...
  /api/v1/teacher/checkin:
    post:
      tags:
        - TeacherPortal
      summary: Giáo viên check-in bằng mã 6 số
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  example: '482913'
      responses:
        '200':
          description: OK
...
  /api/v1/reviews:
    get:
      tags:
        - Reviews
      summary: Đánh giá của phụ huynh (chờ duyệt)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - pending
              - approved
              - rejected
      responses:
        '200':
          description: OK
  /api/v1/reviews/{id}/approve:
    post:
      tags:
        - Reviews
      summary: Duyệt đánh giá (hiển thị công khai)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/v1/reviews/{id}/reject:
    post:
      tags:
        - Reviews
      summary: Từ chối đánh giá
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/v1/reviews/{id}:
    delete:
      tags:
        - Reviews
      summary: Xóa đánh giá
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
...
  /api/v1/parent/reviews:
    get:
      tags:
        - ParentPortal
      summary: Đánh giá của tôi
      responses:
        '200':
          description: OK
    post:
      tags:
        - ParentPortal
      summary: Gửi đánh giá (1-5 sao)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - rating
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 5
                comment:
                  type: string
      responses:
        '201':
          description: Đã gửi (chờ duyệt)
...
  /api/v1/parent/quiz/attempts/{attemptId}/review:
    get:
      tags:
        - ParentPortal
      summary: Xem lại chi tiết 1 lượt làm (đúng/sai từng câu; câu essay kèm essay_score đã chấm tay)
      parameters:
        - in: path
          name: attemptId
          required: true
          schema:
            type: integer
        - in: query
          name: student_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
...
  /api/v1/public/trials:
    post:
      tags:
        - Public
      summary: Đăng ký học thử
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - phone
              properties:
                name:
                  type: string
                phone:
                  type: string
                class_id:
                  type: integer
                  nullable: true
      responses:
        '201':
          description: Đã nhận

  # ------------------------- Phân quyền (roles) -------------------------
...